LottaBot/layer_2/requirements.txt
pandas==2.0.2
aiogram==2.9.2
cachetools
python-dotenv